Complete JWT With Project Structure (16-04-23)

+++ src/main/java/com/hridoykrisna/me/JWTSecurity/project/repository/ProductRepo.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepo extends JpaRepository<Product, Long> {
+public interface ProductRepo extends JpaRepository<Product, Integer> {
 
-    Product findByIdAndIsActiveTrue(Long id);
+    Product findByIdAndIsActiveTrue(int id);
     List<Product> findAllByIsActiveTrue();
 
 }
